+++ BE/src/main/java/com/codesquad/issue/controller/UserController.java
@@ -2,6 +2,7 @@ package com.codesquad.issue.controller;
 
 import com.codesquad.issue.domain.github.GithubAccessToken;
 import com.codesquad.issue.domain.user.UserResponse;
+import com.codesquad.issue.service.OAuthLoginService;
 import com.codesquad.issue.service.UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpHeaders;
@@ -15,17 +16,18 @@ import org.springframework.web.bind.annotation.RestController;
 @RestController
 public class UserController {
 
+    private final OAuthLoginService oAuthLoginService;
     private final UserService userService;
 
     @GetMapping("github-login")
     public ResponseEntity<HttpHeaders> githubLogin() {
-        HttpHeaders headers = userService.redirectGithub();
-        return new ResponseEntity<>(headers, HttpStatus.SEE_OTHER);
+        HttpHeaders headers = oAuthLoginService.redirectGithub();
+        return new ResponseEntity<>(headers, HttpStatus.FOUND);
     }
 
     @GetMapping("login")
     public ResponseEntity<UserResponse> login(@RequestParam("code") String code) {
-        GithubAccessToken token = userService.getAccessTokenByCode(code);
+        GithubAccessToken token = oAuthLoginService.getAccessTokenByCode(code);
 
         return new ResponseEntity<>(userService.userLogin(token), HttpStatus.OK);
     }
