+++ BE/src/main/java/com/codesquad/issue/global/error/GlobalExceptionHandler.java
@@ -1,0 +1,7 @@
+package com.codesquad.issue.global.error;
+
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+
+@RestControllerAdvice
+public class GlobalExceptionHandler {
+}
